Split helper functions into separate testing base class
