+++ test/Makefile.am
@@ -144,7 +144,8 @@ interface_driver_SOURCES = interface/interface_driver.C \
                            interface/antioch_mixture_builder_test.C \
                            interface/absorption_coeff_testing.h \
                            interface/antioch_parsing_type.C \
-                           interface/antioch_build_species_names.C
+                           interface/antioch_build_species_names.C \
+                           interface/rayfire_test_base.h
